Search in Rotated Sorted Array
The problem is to implement a search in O(log(N)) time that gives an idea to use a binary search.

To understand how to find element in sorted rotated array, we must understand first, what is a sorted rotated array? An array is called sorted where for all i and j such that i < j, A[i] <= A[j]. A rotation happens when last element of array is push at the start and all elements on array move right by one position. This is called as rotation by 1. If new last element is also pushed to start again all elements are moved to right again, it’s rotation by 2 and so on.

Find element in sorted rotated array
3 5 6 7 8 9 10 11 15 17
Sorted array
10 11 15 17 3 5 6 7 8 9


Sorted rotated array